CMR-8424: Consolidating function behavior, renaming function to match test name
Cleaning up the verifyEDLJwt script with better naming conventions and having the functionality of different functions take place in only a single function
